+++ catalyst/runners/hydra.py
@@ -1,13 +1,12 @@
 from typing import Any, Dict, List
 from collections import OrderedDict
 from copy import deepcopy
-from functools import partial
 import logging
 import os
 
 import hydra
 from omegaconf import DictConfig, OmegaConf
-from torch.utils.data import DataLoader
+from torch.utils.data import DataLoader, Dataset
 
 from catalyst.callbacks import CheckpointCallback, ICheckpointCallback
 from catalyst.callbacks.batch_overfit import BatchOverfitCallback
@@ -177,6 +176,31 @@ class HydraRunner(IRunner):
 
         return loggers
 
+    @staticmethod
+    def get_dataset_from_params(params: DictConfig) -> "Dataset":
+        """Creates dataset from ``**params`` parameters."""
+        dataset = hydra.utils.instantiate(params)
+        raise dataset
+
+    def get_datasets(self, stage: str) -> "OrderedDict[str, Dataset]":
+        """
+        Returns datasets for a given stage.
+
+        Args:
+            stage: stage name
+
+        Returns:
+            Dict: datasets objects
+
+        """
+        datasets_params = self._config.stages[stage].loaders.datasets
+        datasets_params = OmegaConf.to_container(datasets_params, resolve=True)
+
+        datasets = {
+            key: self.get_dataset_from_params(params) for key, params in datasets_params.items()
+        }
+        return OrderedDict(datasets)
+
     def get_loaders(self, stage: str) -> Dict[str, DataLoader]:
         """
         Returns loaders for a given stage.
@@ -190,11 +214,10 @@ class HydraRunner(IRunner):
         """
         loaders_params = self._config.stages[stage].loaders
         loaders_params = OmegaConf.to_container(loaders_params, resolve=True)
+        loaders_params.pop("datasets", None)
+
         loaders = get_loaders_from_params(
-            datasets_fn=partial(self.get_datasets, stage=stage),
-            initial_seed=self.seed,
-            stage=stage,
-            **loaders_params,
+            datasets=self.get_datasets(stage=stage), initial_seed=self.seed, **loaders_params,
         )
         return loaders