+++ catalyst/utils/data.py
@@ -1,4 +1,4 @@
-from typing import Any, Callable, Dict, Iterable
+from typing import Any, Callable, Dict, Iterable, Union
 from collections import OrderedDict
 import copy
 from functools import partial
@@ -82,9 +82,8 @@ def get_loaders_from_params(
     per_gpu_scaling: bool = False,
     loaders_params: Dict[str, Any] = None,
     samplers_params: Dict[str, Any] = None,
+    datasets: "OrderedDict[str, Union[Dataset, dict]]" = None,
     initial_seed: int = 42,
-    datasets_fn: Callable = None,
-    **data_params,
 ) -> "OrderedDict[str, DataLoader]":
     """
     Creates pytorch dataloaders from datasets and additional parameters.
@@ -98,22 +97,18 @@ def get_loaders_from_params(
             from ``torch.utils.data.DataLoader``
         per_gpu_scaling: boolean flag,
             if ``True``, scales batch_size in proportion to the number of GPUs
-        loaders_params (Dict[str, Any]): additional loaders parameters
-        samplers_params (Dict[str, Any]): additional sampler parameters
+        loaders_params: additional loaders parameters
+        samplers_params: additional sampler parameters
         initial_seed: initial seed for ``torch.utils.data.DataLoader``
             workers
-        datasets_fn(Callable): callable function to get dictionary with
-            ``torch.utils.data.Datasets``
-        **data_params: additional data parameters
-            or dictionary with ``torch.utils.data.Datasets`` to use for
-            pytorch dataloaders creation
+        datasets: ordered dictionary with ``torch.utils.data.Dataset``
 
     Returns:
         OrderedDict[str, DataLoader]: dictionary with
             ``torch.utils.data.DataLoader``
 
     Raises:
-        NotImplementedError: if datasource is out of `Dataset` or dict
+        NotImplementedError: if datasource is out of ``Dataset`` or dict
         ValueError: if batch_sampler option is mutually
             exclusive with distributed
     """
@@ -129,14 +124,10 @@ def get_loaders_from_params(
     assert isinstance(
         samplers_params, dict
     ), f"`samplers_params` should be a Dict. Got: {samplers_params}"
+    datasets = datasets if datasets is not None else {}
 
     distributed_rank = get_rank()
     distributed = distributed_rank > -1
-
-    if datasets_fn is not None:
-        datasets = datasets_fn(**data_params)
-    else:
-        datasets = dict(**data_params)
 
     loaders = OrderedDict()
     for name, datasource in datasets.items():  # noqa: WPS426