+++ CHANGELOG.md
@@ -35,7 +35,26 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
   - Hydra hotfix due to `1.1.0` version changes
 
 
-## [21.05] - YYYY-MM-DD
+## [21.06] - YYYY-MM-DD
+
+### Added
+
+- `dataset_from_params` support in config API ([#1231](https://github.com/catalyst-team/catalyst/pull/1231))
+
+### Changed
+
+-
+
+### Removed
+
+-
+
+### Fixed
+
+-
+
+
+## [21.05] - 2021-05-31
 
 ### Added